+++ frontend/src/pages/PreferRestaurant.tsx
@@ -3,48 +3,20 @@ import PreferRestaurantList from '@domains/restaurant/components/PreferRestauran
 import Button from '@components/actions/Button';
 
 import { PreferRestaurantProvider } from '@domains/restaurant/context/PreferRestaurantProvider';
+import useParticipant from '@domains/restaurant/hooks/useParticipant';
 
 import { apiClient } from '@apis/apiClient';
-import { ParticipantsResponse } from '@apis/prefer';
 import styled from '@emotion/styled';
-import { useState, useEffect } from 'react';
 
 const PreferRestaurant = () => {
   const roomCode = '36f41043-01a3-401d-bdc6-e984b62722d3';
-
-  const [participant, setParticipant] = useState<ParticipantsResponse>({
-    totalParticipants: 0,
-    eliminatedParticipants: 0,
-  });
 
   const handleDeactivate = async () => {
     await apiClient.patch(`rooms/${roomCode}/deactivate`, undefined, {
       'Content-Type': 'application/json',
     });
   };
-
-  useEffect(() => {
-    let isUnmounted = false;
-
-    const fetchParticipantState = async () => {
-      const response = await apiClient.get<ParticipantsResponse>(
-        `rooms/${roomCode}/participants/state`
-      );
-
-      if (!isUnmounted && response) {
-        setParticipant(response);
-      }
-    };
-
-    fetchParticipantState();
-
-    const intervalId = setInterval(fetchParticipantState, 10000);
-
-    return () => {
-      isUnmounted = true;
-      clearInterval(intervalId);
-    };
-  }, []);
+  const { participant } = useParticipant();
 
   return (
     <PreferRestaurantProvider>
